refactor(erp): §S60 — physically split crud_overlay.js's pure CORE into crud_core.js
Implementing bim-compiler prompts/SCRIPT_LENGTH_REFACTOR_SEAMS.md §S60 item 1 (survey 2026-08-21;
extraction precedent: viewer/gantt_model.js, PR #1446).

crud_overlay.js already implemented the pure/DOM split INTERNALLY (var CORE + the node early-return
module.exports = CORE; return) — but ~230 pure lines (sessionActor/changeLog/recordInfo/fieldLineage/
the draft family/_ciKey) physically sat inside the DOM half, reaching node only via function hoisting
across that return. Now the physical layout matches the logical seam:

- erp/crud_core.js (new, 999 lines): the contiguous pure head (old :14-714), the stranded pure block
  (old :2148-2381), and the CORE object literal — ALL moved byte-verbatim (verified programmatically),
  comments included; globalThis.__idmpDb typeof-guard in _getTableCols kept exactly as-is. Dual-mode
  export (window.CrudCore + module.exports). Deliberate deviation from gantt_model's footer: keeps
  crud_overlay's own 'window : this' IIFE binding so node behaviour (global.TipFold / global.APP
  reads) is preserved exactly. Deviation from the spec's function list, with reason: sessionActor/
  sessionClientId/sessionOrgId (old :2148-2161) move too — buildOp :227-229/:244 calls them, so the
  moved head cannot resolve without them; they are typeof-guarded identity READS, same class as adEval.
- erp/crud_overlay.js (2849 → 1921): requires ./crud_core.js in node and re-exports the SAME object —
  require('../crud_overlay.js') is unchanged for every existing caller (identity-checked) — and reads
  window.CrudCore in-browser (fail-loud §-line if load order is wrong). The DOM half is byte-verbatim;
  a small alias block rebinds the names it calls bare.
- glassbowl.html / idempiere.html: <script src=crud_core.js?v=1> added BEFORE crud_overlay (its ?v
  bumped 6→7 / 20→21).
- erp/sw.js: crud_core.js added to PRECACHE_ASSETS + CACHE_VERSION v766→v767 (same-PR rule).

Witness proof (before AND after, logs read): poc_ad_folded_crud 14/14 · poc_doc_dots ALL PASS ·
poc_a_grail ALL PASS · witness_t7_incremental ALL PASS · poc_draft_restore_live 10/10 (live browser —
proves idempiere.html load-order wiring) · poc_critic_odoo_process_live 4/7 and
poc_critic_process_signed_live 5/18 FAIL IDENTICALLY before and after (pre-existing on origin/main;
red-line sets diff-identical). Plus a glassbowl.html wiring smoke: CrudCore present, __crud mounts,
__crud.core === window.CrudCore, 0 pageerrors (5/5).
